Split monomorphize type utilities
