+++ src/typechecker/mod.rs
@@ -22,6 +22,7 @@ mod generic_type_reference_walker;
 mod generic_type_validation;
 mod monomorphize;
 mod monomorphize_inference;
+mod monomorphize_types;
 mod patterns;
 mod program_checking;
 mod resolve;
